fix: acumulado corregido a uint64 (4 registros 0x0018-0x001B)
- config.py: agrega REG_ACUM_W1/W0 (0x001A, 0x001B) para completar los 4 words
- lector.py: lectura de bloque 7 registros, decodificación >Q (uint64 big-endian)
- lector.py: agrega resetear_acumulado() para escritura de ceros vía Modbus (R/W)
- influx.py: flujo_acumulado tipado como int
- README: tabla corregida con 4 registros y columna de acceso R/W

+++ influx.py
@@ -23,7 +23,7 @@ def _get_client():
     return _client, _write_api, _query_api
 
 
-def escribir(flujo: float, temperatura: int, flujo_acumulado: float):
+def escribir(flujo: float, temperatura: int, flujo_acumulado: int):
     _, write_api, _ = _get_client()
     point = (
         Point("caudal")
